Implements namespaced store
